requirements.txt
# -*- coding: utf-8 -*-
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
pydantic==2.10.3
python-multipart==0.0.17
pyyaml==6.0.2
requests==2.32.3
httpx==0.28.1
aiohttp==3.11.10
beautifulsoup4==4.12.3
lxml==5.3.0
python-docx==1.1.2
weasyprint==62.3
pillow==11.0.0
redis==5.2.1
openai==1.57.4
anthropic==0.38.0
playwright==1.49.0
